+++ src/Config/Option/UrlShortener/AutoResolveTitlesConfigOption.php
@@ -17,7 +17,7 @@ class AutoResolveTitlesConfigOption extends BaseConfigOption
     public function ask(StyleInterface $io, array $currentOptions): bool
     {
         return $io->confirm(
-            'Do you want Shlink to resolve the short URL title based on the long URL \'s title tag (if any)? '
+            'Do you want Shlink to resolve the short URL title based on the long URL\'s title tag (if any)? '
                 . 'Otherwise, it will be kept empty unless explicitly provided.',
             false,
         );